Animation support for _anim.txt files
Now we support a card to have an Image as animation when we play it.
The ascii art name must be the same as the card name, just add _anim at the end of the .txt file.
Animation images have a 3 second default to show as soon as they are played.